Moved wrappers function
